refactor: separate section classes in settings
