+++ pubspec.yaml
@@ -19,6 +19,7 @@ dependencies:
   http: ^1.1.0
   path: ^1.8.3
   intl: ^0.18.0
+  tuple: ^2.0.2
   win32: ^5.0.5
   async: ^2.11.0
   shimmer: ^2.0.0
